Add random tick intervals to try to match nohz
Since NOHZ is likely to be fairly random in its tick distribution, I
figured the simulator should consider this.

So here's a patch to add infrastructure for get_random_int(), and to use
it for picking random tick intervals for each sample.

For now use a constant seed so we get reproducible results.

Signed-off-by: John Stultz <john.stultz@linaro.org>